+++ R/pmcmc_tools.R
@@ -69,6 +69,9 @@ pmcmc_filter <- function(object, i) {
   if (!is.null(object$trajectories)) {
     object$trajectories$state <- object$trajectories$state[, i, , drop = FALSE]
   }
+  if (!is.null(object$restart)) {
+    object$restart$state <- object$restart$state[, i, , drop = FALSE]
+  }
   object
 }
 
@@ -131,7 +134,17 @@ pmcmc_combine <- function(..., samples = list(...)) {
   if (is.null(trajectories[[1]])) {
     trajectories <- NULL
   } else {
-    trajectories <- combine_trajectories(trajectories)
+    trajectories <- combine_state(trajectories)
+  }
+
+  restart <- lapply(samples, "[[", "restart")
+  if (!all_or_none(vlapply(restart, is.null))) {
+    stop("If 'restart' is present for any samples, it must be present for all")
+  }
+  if (is.null(restart[[1]])) {
+    restart <- NULL
+  } else {
+    restart <- combine_state(restart)
   }
 
   ## Use the last state for predict as that will probably have most
@@ -140,27 +153,25 @@ pmcmc_combine <- function(..., samples = list(...)) {
   ## We might check index, rate and step here though.
   predict <- last(samples)$predict
 
-  mcstate_pmcmc(pars, probabilities, state, trajectories, predict,
-                chain, iteration)
+  mcstate_pmcmc(pars, probabilities, state, trajectories, restart,
+                predict, chain, iteration)
 }
 
 
-combine_trajectories <- function(trajectories) {
-  base <- lapply(trajectories, function(x) x[names(x) != "state"])
+combine_state <- function(x) {
+  base <- lapply(x, function(el) el[names(el) != "state"])
   if (length(unique(base)) != 1L) {
-    stop("trajectories data is inconsistent")
+    stop(sprintf("%s data is inconsistent", deparse(substitute(x))))
   }
 
-  ## This is nasty:
-  trajectories_state <- lapply(trajectories, function(x)
-    aperm(x$state, c(1, 3, 2)))
-  trajectories_state <- array(
-    unlist(trajectories_state),
-    dim(trajectories_state[[1]]) * c(1, 1, length(trajectories)))
-  state <- aperm(trajectories_state, c(1, 3, 2))
-  rownames(state) <- rownames(trajectories[[1]]$state)
+  state <- lapply(x, function(el) aperm(el$state, c(1, 3, 2)))
+  state <- array(
+    unlist(state),
+    dim(state[[1]]) * c(1, 1, length(x)))
+  state <- aperm(state, c(1, 3, 2))
+  rownames(state) <- rownames(x[[1]]$state)
 
-  ret <- trajectories[[1]]
+  ret <- x[[1]]
   ret$state <- state
   ret
 }