+++ NEWS.md
@@ -1,3 +1,8 @@
+# mcstate 0.3.1
+
+* The particle filter can now return the entire model state at points during the run, with argument `save_restart` to `$run()` and method `$restart_state()` (#86)
+* The `pmcmc` can returned sample restart state using the `save_restart` argument to `mcstate::pmcmc_control` which can be used to restart the pMCMC part way along the time series (see `vignette("restart")`)
+
 # mcstate 0.3.0
 
 * `pmcmc` is now controllable via a new `mcstate::pmcmc_control` object